tinyspeech/pipeline.py
# TinySpeech Summarizer - Core Pipeline
# This file contains all functions for transcription, summarization, and file output

import os
import whisper
import google.generativeai as genai
from pydub import AudioSegment
from datetime import datetime
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

def load_whisper_model():
    """Load the Whisper tiny.en model for transcription"""
    # TODO: Implement Whisper model loading
    pass

def transcribe_audio(audio_path):
    """Transcribe audio file using Whisper"""
    # TODO: Implement audio transcription
    pass

def generate_summary(transcript):
    """Generate summary using Gemini API"""
    # TODO: Implement Gemini API summarization
    pass

def create_markdown_report(transcript, summary, chapters, audio_filename):
    """Create a formatted Markdown report"""
    # TODO: Implement Markdown report generation
    pass

def process_audio(audio_path):
    """Main pipeline function to process audio file"""
    # TODO: Implement complete audio processing pipeline
    pass

if __name__ == "__main__":
    # CLI interface for testing
    import sys
    if len(sys.argv) > 1:
        audio_file = sys.argv[1]
        process_audio(audio_file)
    else:
        print("Usage: python pipeline.py <audio_file_path>")
